Regression issue - restore working code in StandAloneProcessProxy
The previous changes to add support for pulling connection files
introduced a bug in the StandAloneProcessProxy in order to properly
test pulls.  That change broke the previous push model (which is what
we want in that particular proxy) by not setting up the remote IP
correctly (left at 0.0.0.0).  This change restores the code to working
order.