Detect older versions of notebook that do not contain culling feature.
If older versions of notebook (< 5.1.0) installed, the idle kernel
culling functionality is not present.  This can lead to a failure to
start Elyra in cases where Elyra was shutdown with active kernels.  In
those cases, Elyra's restart will try to activate those kernels and, if
they can be activated, the idle kernel culler initialization is called.
This call then produces an AttributeError exception if an older version
of notebook is present.

Closes #46