Make sure connection file is still produced when pull mode is enabled.
Other changes.
1. Added checks that if the connection file already exists, then we'll
assume 'push' mode.
2. Added some validation of the response address parameter since users
may forget to remove that parameter from the LAUNCH_OPTS variable.  In
such cases, they will encounter ERROR__NO__KERNEL_RESPONSE_ADDRESS
as the response-address parameter value.  If we find that the response
address cannot be split into two pieces on ':' or if the port portion
of that split is not an int(), then assume 'pull' mode has been
specified and just return.