Detect loopback scenario when using socket mode.
Found that we need to remove the local connection file prior to launching
Yarn kernels with pull or socket mode.  This allows the kernel to write
the file locally for loopback and prevents the need for having to rewrite
those files.
Also found that we were attempting to use the file too soon on loopback
because the kernel launcher hadn't written the file when Elyra started
trying to use it. Instead, the rcp code - which detected loopback and
skipped the copy, will throw ENOENT if the file doesn't exist which lets
the existing retry logic work for both remote and loopback files.
Also cleaned up the update_connection method so that the file is only
written when necessary.