Ignore additional interrupt attempts after receiving the first.
When Elyra is started in the foreground, entering control-c from the
keyboard will interrupt the IO loop and trigger the graceful shutdown
of the service - which includes attempts to shutdown any active kernels.

However, if a second interrupt (control-c) is issued during the shutdown,
that interrupt will terminate the graceful shutdown activities taking
place.  This change explicitly ignores additional interrupt signals,
thereby permitting the original shutdown sequence to complete.