Decrease sleep timeout and increase socket timeout.
Decreased the non-interruptable sleep timeout from 1 to .5 seconds so
that we check status of connection information more frequently.

Increased interruptable socket timeout from 1 to 5 seconds so that we
optimize socket mode operations since the method will be interrupted
when the data arrives.  This way we don't spend time waiting on
non-interrupted calls.

Also performed a little more refactoring to cut down on the number of
Yarn queries performed each iteration.