Tolerate missing env when starting kernels
Elyra currently assumes an environment variable (env) exists in the set
of arguments to starting a kernel. However, when starting a kernel
directly via the REST API, env is does not exist. This change establishes
an env if one does not exist and adds KERNEL_ID so that remote kernels
can be launched.

The session persistence stores the value of KERNEL_USERNAME and had been
assuming its existence.  This change also tolerates that value not being
present.

Updated the pool code to use a copy of the environment since that is a
better practice.

Fixes #87
Closes #89