Remove --response-address parameter from kernel.json LAUNCH_OPTS
This change removes the --response-address parameter from the `LAUNCH_OPTS`
env variable specified in the kernel.json.  Instead, whenever `socket`
mode is used, Elyra will modify the kernel run command to include the
parameter.

The `LAUNCH_OPTS` variable will remain in the kernel.json file but will
default to the empty string, since this will still allow users to
influence their launcher if necessary.

This change is a precursor to Issues #80, #90 and #91.

Closes #78