Launchers should determine location of connection file if it doesn't exist
In order to reduce installation/configuration steps on worker nodes,
launchers now check if the "proposed" connection file directory (assuming
the file doesn't already exist) exists.  If it doesn't exist, the launcher
is responsible for creating a temporary location and removing the
connection file when the kernel terminates.

All launchers now make this check and produce a message indicating that
the original (proposed) location is changing.  This is performed by the
`determine_connection_file()` method.

Elyra now does NOT write the returned connection information to the
default file location since the file is only needed by the kernel.  As
a result, only the kernel manager member variables are updated and used.
FYI: This is how the load of persisted kernel sessions works.

Also added support for env `ELYRA_KERNEL_SESSION_LOCATION` to allow
setting the kernel session persistence location.  It will default to
the built-in jupyter data directory.  Related to #116.

Fixes #107
Closes #117