Do not auto restart remote kernels without connected clients
Intercept kernel restarts and check if the kernel's process proxy
is an instance of RemoteProcessProxy. If so, check its number of
open connections. If zero, use the parent mapping kernel manager
instance to shutdown the kernel rather than restart it. This will
allow admins to terminate resource-managed (e.g., yarn) applications
of disconnected kernels rather than having to wait the entire culling
period for the resource to be.

Fixes #118
Closes #122