Skip restarts of non-connected kernels when triggered from restarter.
The previous attempt to address issue #118 broke restarts triggered
from the notebook client because the notebook actually closes the
websocket after triggering the restart - causing the connection count
to go to zero.  As a result, it is not sufficient to base this change
on a zero connection count.

Instead, we will use the `now` parameter value.  Its default is `False`
and `True` is only used when called from the restarter (i.e., when
auto-restart is the trigger).