Use SSH tunnels for secure socket communication
Secure communications between Enterprise Gateway and the kernel
has been implemented by SSH tunnels, and it requires passwordless
SSH to be properly configured. While this is enabled by default,
it can be disabled by configuring the envitonment variable
EG_USE_SSH_TUNNELING as False when launching Enterprise Gateway.

This was implemented by @akchinSTC. Fixed minor issues to get it
to work end-to-end.

Fixes #100
Closes #166