Add support for gateway communication port
Because signals cannot be used between processes of differing
user-ids (of the gateway server and kernel process), we need
a different mechanism for communicating to the kernel that it
needs to interrupt its current cell.

This change alters Enterprise Gateway to look for the return of
a comm_port value upon return of the connection information from
the kernel launcher. If present, that port (along with the ip
corresponding to the kernel's assigned host) will be used as a
gateway communication port. Currently, the port listener within
the launcher understands one request. If the json payload contains
a signum key, then its value is the corresponding signal number the
launcher will send to the kernel (which is essentially itself). Likewise,
if the comm_port has been returned, the Enterprise Gateway will attempt
to send corresponding signal values via the gateway communication port.
If an exception occurs, the current fallback mechanism will be attempted,
although this will not succeed if the processes are running under different
user-ids.

The ipython kernel launcher has been updated accordingly. This logic is
enabled by default but can be disabled by invoking the launcher with
--RemoteProcessProxy.disable-gateway-socket. When enabled,the launcher
builds the connection information, then creates the gateway socket for
listening and returns the port in the comm_port key value of the connection
info response. It then starts a thread to listen on the socket, while the
primary thread completes the launch of the kernel. If the listening thread
receives data, it checks is there's a signum key and uses that value to signal
itself - which is sent to the primary thread. If there's no signum key,
the request is ignored (until new requests are needed).

This gateway communication port will be tunneled if ssh tunneling is enabled.
As a result, some refactoring of the recently added tunneling code has occurred,
along with fixing the detection of whether or not ssh tunneling is enabled.

Closes #169