Gracefully shutdown kernel when comm listener is used
When used, the comm port listener ran infinitely, keeping the thread
active.  This interfered with graceful kernel shutdowns since it
essentially kept the launcher alive - thereby leading the gateway
to believe the kernel was still alive - triggering the kill sequence
5 seconds later.

This change introduces another request to the launcher. If the received
data contains a 'shutdown' key then its value is cast into a boolean
(where '1' is the expected value).  The while loop has been modified
to let this value trigger its termination - thereby exiting the thread.

RemoteKernelManager now overrides `request_shutdown()` such that if
there's a remote process proxy in use for the kernel, its
`shutdown_listener()` method will be called.  If a listener is in use
(by virtue of the establishment of a comm_port) then the aforementioned
shutdown request is sent to the listener.

Closes #176