Add support for gateway listener to Toree Launcher
Extended the toree launcher to return the port number corresponding to
a socket that has a thread listening for interrupt and shutdown requests.

Updated the cluster kernel.json for the scala kernel to include the
`SPARK_YARN_USER_ENV` entry that includes the `TOREE_ALTERNATE_SIGINT=USR2`
entry so that the alternate signal indication is set into the toree
launcher/kernel's environment - otherwise the interrupt will terminate
the kernel - since toree doesn't know to honor that value and setup a
handler.

Also added better protection around the `pid`, `gpid` and `comm_port`
return values from the launchers so any non-integer values are essentially
ignored (with applicable warning messages).

Fixes #170
Closes #182