Skip test_default_kernel_name until Notebook issue is resolved.
A recent change to Notebook (PR 2853) introduced an issue such that
the NoSuchKernel string is no longer present in the response body
when attempting to use a non-existent kernel.  This condition is
the basis for checking of the default kernel is used.  As a result,
that test is now marked as skipped until that Notebook issue (2942)
is resolved or addressed.