Remove use remote-user command line option and env variable.
We no longer document the use of `EG_REMOTE_USER` or
`--EnterpriseGatewayApp.remote_user` but instead that the _service user_
needs to be configured for password-less ssh.

In the code, the use of `EG_REMOTE_USER` and optional `EG_REMOTE_PWD`
is still present, but remote user defaults to `$USER` and password defaults
to `None`.  These merely exist in case we need to leverage this capability.

Added another example (`SSHException`) to the troubleshooting instance for
password-less ssh issues.

Fixes #198