Eliminated sleep to keep SSH session alive
Earlier, sleep maxint  was being used to keep the
SSH session alive. However, even after the kernel
process was closed/halted and the SSH processes were
terminated, the  child processes would not
get terminated. As a result, the number of
child processes running under the service-user would
keep increasing as more and more kernels were launched
on the remote host.

This change eliminates the sleep child process as we
no longer need it to keep the SSH session alive.

Also, no longer passing in timeout to the paramiko_tunnel()
method. However, I am keeping references to maxint as it
will be used for other features such as keepalive.