Specify ServerAliveInterval on SSH sessions
In order to ensure that SSH connections stay alive longer
than the Kernels, Enterprise Gateway now explicitly
specifies ServerAliveInterval option on the SSH sessions.

For CONTROL and COMMUNICATION channels, SeverAliveInterval
is currently hardcoded to the number of seconds in 100years.
For other(SHELL, IOPUB, STDIN, and HEARTBEAT) channels, the
value of ServerAliveInterval is computed by adding 60 seconds
to the value of the MappingKernelManager.cull_idle_timeout
command-line parameter. If MappingKernelManager.cull_idle_timeout
is unspecified or beyond acceptable range, then the channels
get treated like CONTROL and COMMUNICATION channels with respect
to ServerAliveInterval value.

Fixes #207
Closes #211