Add toree-launcher to enterprise-gateway repo
Moved the toree-launcher from a private repo into EG. As a result, performing
`make kernelspecs` now requires the installation of `sbt` - which will build
the toree launcher jar and place it into the `etc/kernel-launchers/scala/lib`
directory (where it was located prior to this change).  From there, the build
then includes it when archiving the kernelspecs into a tar file.

The file name for the torre-launcher has been altered from `toree-launcher_2.11-1.0.jar`
to `toree-launcher_2.11-0.7.0.SNAPSHOT.jar` (or whatever EG version is configured).
If this is not a `dev[0-9]` build, then `SNAPSHOT` will not be included in the filename.

Since there is now a `Makefile` in the `etc` directory to logically split the
building of the kernelspecs to their repo location, the similar split has taken
place for the docker images.  All targets are still accepted by the primary `Makefile`,
just that the `kernelspecs` and `docker *` targets get propagated to `etc/Makefile`.

Fixes #223