Remove spark version indicator from kernelspecs
Removed references to the spark version (`2.1`) from the kernelspecs (display
names and echo statements), renamed directories accordingly, and updated like
references in the docs.

Also, updated the doc reference to the `0.6.0` version of the kernelspecs tar
file to `0.7.0` since we will be cutting a release later this week.  As a result,
there may be a brief window where that corresponding link is invalid.

Fixes #174
Closes #227