[HOTFIX] Fix broken link to kernelspecs tar file. (#230)
Fixed the broken link to kernelspec tar file. However, while on this
page I discovered the need for the following changes.

1. We had duplicate statements for Scala and Python sections so I moved
to the beginning section where all three kernels are referenced.
2. Both code blocks for Scala and Python instructions referenced invalid
links to 0.6 and then used referenced invalid tar files.  With the changes,
they *should* be able to cut/paste the instructions.
3. We were missing the comparable section for the iR kernel - so I added it.

Fixes #229