Add JupyterLab option to elyra/nb2kg image.
Now that nb2kg can run with notebooks > 5.0, we can extend the docker
image to include the JupyterLab extension.  This pull request does the
following:

- Removes the requirement that a particular docker image (corresponding
to a pre-5.0 Notebook version) be used.  Now the latest minimal-notebook
image is used.
- Installs the Jupyter Lab server extension.
- Updates the entrypoint to recognize a `lab` command.  If the container
is started with `lab` as the `CMD`, it enables the jupyter lab server
extension and invokes `jupyter lab` with the `NB2KG` extension.
- Defaults the `CMD` to `notebook` such that no command or a command of
`notebook` launches `jupyter notebook` with the `NB2KG` extension.
- Any `CMD` values that are not `lab` or `notebook` will invoke the bash
shell into the container (for troubleshooting).
- Updates the README accordingly.

Fixes #236
Closes #237