Add proxy configuration to kernelspec
With this PR, administrators have the ability to set
configuration values relative to specific kernels via
the `process_proxy` stanza. In addition, some configuration
values can be used to override system-level values, thereby
allowing administrators to better control/segment system
resources or user authorizations.

The proxy configuration dictionary is read during the
load of the kernelspec and conveyed to each process-proxy
constructor in the given class hierarchy. Which classes utilize
the information is a function of that particular class'
purpose.

See online docs for details and examples.

Fixes #161
Closes #240