Fix SSH tunnel termination on kernel shutdown
This PR does two things as discussed in Issue #243.

1. In _forced kill_ situations (when the kernel is determined active after the
initial shutdown request), the `cleanup()` method on the process proxy classes
was not getting called.  This is because the `cleanup()` method override was
checking if the `self.kernel` variable was set.  However, because the `_kill_kernel()`
method (that is triggered in forced kill situations) sets the `self.kernel`
member to None, the subsquent call to `cleanup()` resulted in the process proxy
cleanup not being called.

Since `self.process_proxy` is the same value as `self.kernel`, the `cleanup()`
override checks if that member is not None and calls `self.process_proxy.cleanup()`
leading to the termination of SSH tunnel processes - among other things unrelated
to SSH tunneling.

2. The EG server believes the kernel/launcher is still active after the initial
shutdown request, even though the kernel/launcher is terminated because the
corresponding SSH tunnel is still constructed.  As a result, the framework's
forced kill sequence is triggered.  By including the termination of the
communication port's SSH tunnel following the request to shutdown the listener,
subsequent `poll()` requests now fail resulting in the avoidance of the forced
kill sequence.

Fixes #243