Validate cell execution results immediately after execution
With this change, cell execution results are validated immediately
following the cell's execution.  This allows the output to better
convey what is going on in the test.  Output is prefixed with `>>> `
and an indication of whether the execution is considered passed
or failed is also displayed.

The flag --continue_when_error is now properly honored.

Fixes #247
Closes #249