Enable SSH tunneling by default (#250)
For enhanced security, the connections between the
Jupyter Enterprise Gateway and the Kernel as well as
the connection between the Enterprise Gateway and the
Kernel Launcher have been made secure by default. The
documentation has been updated as well with details
about how to turn off tunneling for troubleshooting
purposes.