Fix signal_kernel() to nolonger assume its only used for interrupts
A recent change to jupyter_client started using the `signal_kernel()` method
on `KernelManager` for signals other than `SIGINT`.  However, the override
method in `RemoteKernelManager` erroneously assumed that was only called for
interrupt scenarios.  As a result, there was a path that allowed an unset
variable to be used as the signal number rather than the value passed in.

This change fixes that scenario and confirms that tests which were previously
failing when jupyter_client 5.2.0 is installed, now succeed.

Fixes #251