Toree Launcher: --alternate-sigint command-line option (#259)
Replaced the use of environment variable TOREE_ALTERNATE_SIGINT
with a command-line option --alternate-sigint that is passed in
using __TOREE_OPTS__ in the kernel.json files. Updated the
relevant kernel.json files as well. Also, removed dead code in
remotemanager.py that was dealing with EG_ALTERNATE_SIGINT
environment variable.