Add support for kernel interrupts for R kernels
Found that the `system()` command can be used with a `wait=FALSE`
option which essentially allows for asynchronous behavior.  As a
result, by combining the python code to create a thread to listen
on the communication port with the code that writes out the
connection file, we can essentially get a thread that then forwards
SIGINT signals to the IRkernel and interrupt a cell's execution.

This required an update to the yarn-cluster run.sh script so that
the co-located python file (gateway_listener.py) is propagated in
cluster mode.  For distributed mode, its co-location with the R
launcher on the target node is sufficient.

Fixes #171