Add support for port-range to the Enterprise Gateway server
Add formal command line option `--EnterpriseGatewayApp.port_range` to EG
(default value = 0..0, ranges of zero disable functionality).  This value
can also be overridden on a per kernelspec basis via the "config" stanza:
```
    "config": {
        "port_range": "33333..34333"
    }
```
The configured range can be conveyed to the kernel via the "argv" stanza
entry:
```
    "--port_range",
    "{port_range}"
```
With this change, any port created on the EG server (e.g., the response-
address port and all tunnel endpoint ports) will also adhere to the active
port-range values.

Added code to detect that when a remote kernel is launched and is using
a response-address, then the connection file on the EG is not written.
This prevents the creation (and closure) of 5 useless ports that would
not adhere to port-range restrictions and, potentially, could generate
failures (depending on the configured restrictions).