Make enterprise-gateway image more usable
The enterprise-gateway image can be used in a combined mode or dual mode (where
one instance runs as the YARN cluster and the other just EG).  The yarn-spark
image is also "runnable" as the YARN cluster to make it more clear to users
using dual mode.

All of the env variables can be optionally set via the docker command line
and there's now a mount point of `/tmp/byok/kernels` that allows users to
"bring their own kernels".  This, combined with the ability to specify a
remote YARN host, should enable use of the enterprise-gateway image to
target other YARN clusters.

The docker hub repos have been updated accordingly at:
https://hub.docker.com/r/elyra/enterprise-gateway/
https://hub.docker.com/r/elyra/yarn-spark/

One of the recent changes to make SSH more secure is side-affecting the
DistributedProcessProxy kernels (YARN client mode), but SSH is not used for
YarnClusterProcessProxy kernels (YARN cluster mode).

Fixes #267