RFC 793 compliance for port range
Since port number is a 16-bit unsigned int, it must be in
the range (0, 65535). However, TCP also sets aside ports in
the range (0, 1024] as well-known ports that are reserved
for specific purposes. This change set adds checks to ensure
that specified port range falls within (1024, 65535). Updated
all the kernel.json files by adding the argument
--RemoteProcessProxy.port-range. Updated the configuration
and troubleshooting documentation.

Fixes #290
Closes #303