Add support for enforcing max number of active kernels per user.
Added option `--EnterpriseGatewayApp.max_kernels_per_user`
(env: `EG_MAX_KERNELS_PER_USER`). Default value is -1 and disables
enforcement.  When >= 0, the sessions dictionary is consulted
to determine if the current number of sessions for the given user
is less than the imposed maximum. If the current count is >= the
imposed maximum, a 403 will be raised with an appropraitely logged
message.  Otherwise, kernel satartup will proceed and the new session
will be tracked.  Kernel shutdown removes the previously tracked session.

This change enables the in-memory session management logic that had been
conditionally enabled.  The persistence logic is still disabled by default.

Fixes #295
Closes #306