Save original connection information when tunneling is enabled
When persistent sessions are enabled and tunneling is in use, the
tunnels need to be rebuilt when loading persisted sessions from storage.
This change refactors some of the existing tunnel creation logic so
that it properly captures the "tunneled" information and can also be
called when loading process info during startup.

Fixes #167
Closes #314