Remove code that terminates launcher prematurely
The launcher was terminating itself on shutdown requests when it
should terminate its listener loop.  This was resulting in KILLED
applications on Conductor (and YARN).  Note however, that after
this change, while Conductor applications are gracefully terminated
and indicate FINISHED, YARN applications still appear as KILLED.
We'll need to look into why YARN applications are still not shutting
down gracefully.

Fixes #317
Closes #318