Fix LocalProcessProxy kernel restarts
Kernel restarts were timing out when `LocalProcessProxy` (default)
kernels were in use.  This was due to a change made in v0.8.0 to
support port ranges on local kernels.  The issue was caused by
new port sets getting generated unnecessarily.  As a result, the
restart sequence was messed up, leading to timeouts.

By skipping this area when restarts are in play, restarts can
happen for all configurations.