Add support for initialization modes
This PR adds support to spark context intialization modes to the python
and scala kernel launchers.  PR #324 added this support to the R kernel
launcher.

- Python launcher: Since the python launcher already performs automatic
spark context creation in the background, modes `lazy` and `eager` behave
the same. Some minor refactoring was required to support `none` so that
the pyspark.sql package wasn't loaded.

- Scala launcher: Support for the initialization modes are provided by
the Toree kernel in either `--spark-context-initialization-mode` (for
lazy and eager) or `--nosparkcontext` (for none).  Since we don't want
to have to set both the launcher and Toree options in the kernel.json
some more significant rework of the arguments processing was required.
This also addressed an issue regarding the support of flag-based
parameters.  Processing of the `--alternate-sigint` parameter was also
refactored such that its processing takes place with all other arguments.

The applicable kernel.json "samples" were updated to relfect the `lazy`
option (since that's the default) with the exception of the Conductor
python file - which wants to use `eager` (although its the same behavior
per above).

Fixes #316
Closes #325