Ensure appropriate status codes and messages
Visited each of the areas where we raise `HTTPError` or `RuntimeError`
and made sure the error was appropriately logged and messaged back
to client.  With few exceptions, RuntimeErrors were converted to
HTTPErrors with a status code of 500.

Since raising HTTPError does not produce any entry in the server
log, I introduced a base class method `log_and_raise()` that will
both log and raise the exception.  Depending on the value of the
`http_status_code` parameter, the method will raise an `HTTPError`
or `RuntimeError` (although cases of the latter do not currently
exist).

Most status codes remained as is with conversion of `RuntimeError`
cases (primarily port-range validation) to `HTTPError` 500.  Generally
speaking, EG will raise status codes of `403`, `500` and `503`.

Not all of the locations where exceptions are raised could be tested
but I validated a number of them by temporarily sabotaging my
configuration (primarily via malicious edits to `kernel.json`, `run.sh`,
and launcher scripts).

Fixes #311
Closes #328