Prepare for integration with spark on kubernetes
In order to prepare for integration with spark on kubernetes, the
following changes were made:

- Removed use of Job for kernel. Now, kernels are represented by a Pod.
This is analogous to how spark-drivers are represented and I found that
we didn't really need the Job object.  As a result, all discovery and life
cycle operations are identical, regardless of whether the pod is a vanilla
kernel, or a spark driver.
- The enterprise-gateway container now runs as the elyra user
- The kernel image no longer uses a fixed connection file and ports.  Instead
enterprise gateway conveys the connection file name as an environment variable
and the launcher does exactly what it did before - creating the ports it needs
along with the key value.
- The launcher now optinally creates a spark context.  As a result, we should
be able to use the same launcher for vanilla kernels as for spark.  At this time
we still use a separate launcher in the k8s kernel image, but the idea is that
it will use the same launchers used for spark jobs.