Properly handle max kernel per user on restarts
During restarts, the max kernel per user enforcement wasn't taking into
account that this kernel startup was due to a restart.  As a result, if
the number of current sessions was already at the threshold, the server
would raise an exception - assuming this startup would count against the
active sessions.  Since restarts re-use the kernel, they should not be
considered in this enforcement.

Fixes #331