Detect missing jar files during scala kernel launch
Because scala kernel launch is dependent upon the separately installed
Toree jar file, this PR adds a check to ensure that file exists and, if
not, echos an error message and exits the script.  While at it, we also
ensure that the toree kernel launcher jar is present as well.

Since there isn't a very graceful way to communicate this issue, the
changes also include some improved log statements instructing where
users should look to gain details regarding the launch failure.

Fixes #327