Expose KERNEL_ID and KERNEL_USERNAME to kernel pods for spark AppName usage
Since we don't automatically create a spark context for the kernel pods (with
the exception of Scala/Toree), it would be helpful to provide instance-specific
information like KERNEL_USERNAME and KERNEL_ID to the env of the pod.

Because the Toree Launcher still creates a spark context, use the two env
variables to set the context's appName to ${KERNEL_USERNAME}-${KERNEL_ID}
to be consistent with the cluster startup scenario.