Add ability to specify executor image, use same image for driver and executor
Since users will likely want to add various jars to the images we provide
and since driver and executor images only differ slightly in their startup
commands, it makes sense to use the same image for the driver and executor
and let the CMD determine the target invocation based on env.

This change extends the process-proxy config with `executor_image_name:`
to indicate the image to use for executors. This value will be posted to
the env variable: `EG_KUBERNETES_KERNEL_EXECUTOR_IMAGE` for use in the
spark-submit call.