Add ability to launch kernels is separate namespaces
KERNEL_NAMESPACE is now used to convey in which namespace the kernel
should be launched.  If this variable is not provided, one will be
created and defaulted to the value of EG_KUBERNETES_NAMESPACE. EG
assumes this namespace already exists and its lifecycle is separately
managed.

For spark-on-kubernetes kernels that are to be launched in a
separate namespace from EG, an additional configuration option
(`--conf spark.kubernetes.authenticate.driver.serviceAccountName=<svc-acct>`)
will likely be necessary which denotes the service account that should be
used to create the executors.  This information will be documented.