Ensure message ids match when processing requests
On occasion, all integration tests for a given kernel would fail because
the first response would return the empty string, then all other responses
would be off-by-one from their expected result.  The reason for the skew
appears to be because some messages are returned that do not have the
same message id as the one used during the kernel's creation.  The protocol
is to ensure that responses relative to the same message id correspond to
the kernel. By ensuring that only messages with the same message id are
processed, the skew issue is resolved.

I believe there's an idle status returned where the message id is either
not set or of a non-matching value.  This was resulting in the loop's early
termination with a cell result of the empty string.