Cleanup makefile, add tensorflow-based kernel
Cleaned up the Makefile for building images.  All of the redundant code
has been replaced with two functions to generate build and cleanup targets
for all of the images.  This required removing the string `docker-` from
the 3 plain docker images (enterprise-gateway, nb2kg and yarn-spark).

(We should revisit the names of the kubernetes images once we get more active.)

Also added another kernel image based on the Tensorflow community image.