Add ability to set kernel id
This change allows a client to use a user-specified kernel_id by setting
the `KERNEL_ID` environment variable into the env dictionary of the json
request.  This can be used in conjunction with `KERNEL_NAMESPACE` so that
consumers can have everything based on the kernel id, including the
namespace in which the kernel containers exist.

This change requires that a version of jupyter_client newer than `5.2.3`
be used since this change is implemented via a method override.  Attempts
to set a `KERNEL_ID` using juptyer_client version <= `5.2.3` will result
in the env value being ignored and kernel_id will be system-generated.

The value of the `KERNEL_ID` must be a string-ized uid and version-4
compliant. Any other kinds of values will result in an exception when
starting the kernel.