Changes necessary to build and run docker integration tests on travis
Extended travis configuration to include docker attributes, and use
makefile for running integration tests.

Removed testing for Python 3.3 since it has been deprecated and 3.5
since it's in between 3.4 and 3.6.

Changed assert statements to use self.assertRegexpMatches - which
will then print the result string on failures.

Added -s flag to nosetest so that print statements are displayed.

Reworked logic to use a reader thread that is responsible for reading
all messages from the socket (across the tests for the kernel). This
reader thread then places the messages into queue which is added to
a dictionary based on the msg_id for that execution request. By using
a queue for each execution, we can read past the IDLE status to handle
out of order messages.  Each execution (i.e., test) pulls messages
from its queue.  Once IDLE is encountered, it issues additional reads
with a .5 second timeout to see if there are out-of-order messages,
then terminates the loop.  The reader thread terminates when the kernel
is shutdown.

Note: travis build/test times have now increased from around 5 minutes
to closer to 20 - with about 10 minutes spent building the two docker
images. We'll probably want to try to improve that, especially as
the tests increase in size.

Fixes #293