Restored code to send alternate interrupt signal numbers for local cases
Since EG still supports local kernels (that go through LocalProcessProxy)
without a launcher, we still need to support the ability to specify an
alternate signal number.  As a result, the code to look for `EG_ALTERNATE_SIGINT`
in the kernel.json's `env` stanza has been restored.  When present, this
informs EG that a different signal (the value of the env variable) be sent
instead of `SIGINT`.  In nearly all cases, the value of this variable will
be `SIGUSR2`.

Closes #352