Add restart and interrupt tests for each kernel
Restart tests ensure that a new kernel is started on the server and
verifies that the application_Id associated with the new kernel is
of a greater value than prior to its restart.  Of course, this is
only true for kernels based on YARN.

Interrupt tests must create a thread to perform the interrupt after
the kernel has been issued a long-running statement.  The tests
confirm that an interrupt occurred (different depending on kernel
type) and that a subsquent fetch of the application Id yields the
same value as that prior to the interrupt.

Minor class refactoring in preparation for testing vanilla kernels.

Add KERNEL_LAUNCH_TIMEOUT (default = 40) to kernel start requests.