Fix unauthorized_users list
The authorized and unauthorized users lists are implemented using Sets.
Since sets are immutable, the code was performing a union between the
globally scoped unauthorized_users list with the one defined in the
process proxy. However, it wasn't re-initializing the variable that is
then used downstream.  As a result, the code behaved as if the process
proxy-defined list was not in use.

Since unions are not performed on the authorized_users lists, this issue
is not present there.

Fixes #356