Add initial authorization tests
Added tests to validate the use of the authorized_users and unauthorized_users
lists that can be configured in the process-proxy:config stanza of a kernelspec.

This change also uses local kernelspec definitions that are mounted into the
docker image via the byok (bring-your-own-kernelspec) capabilities recently
added to image.  The idea is that different test-specific kernelspecs can be
added when necessary.

An actual bug (Issue #356) was identified during development of these changes.

Note: The travis builds associated with this PR will fail until PR #357 has
been merged since that contains the fix for the unauthorized_users list that
is tested as part of these changes.