Add initial unittests specific to enterprise-gateway
Extended the TestJupyterWebsocket class with TestEnterpriseGateway.
This class contains three tests:
1. test_max_kernels_per_user - verifies that kernels are limited per
user.
2. test_authorization - is a similar test to the one performed in the
integration but addresses the globally scoped authorized_users and
unauthorized_users lists (where the integration testing validates the
process-proxy override behavior).
3. test_port_range - validates the logic behind port range restrictions.
This test only uses the global setting, but its the same code in the
server once the values have been pulled from the process-proxy config
if overrides are in play.

Closes #359