Use value for PYSPARK_PYTHON as python command to launch listener
In environments where multiple python executables are installed, its
important to use the python env that is also used for Spark. This change
will honor the value of `PYSPARK_PYTHON` if present, then default to
using which `python` is found in the PATH, prior to launching the
background gateway-listener process.

Fixes #362