Update docs to include kubernetes support information
Updated the online documentation to include the necessary information
for Kubernetes support.  This consisted of extending our set of process
proxy implementations and adding descriptions for the various images.

I also updated the elyra dockerhub kubernetes images with the contents
of their respective README.md files (added in this change).

A general walk through of the entire online documentation set was performed
and I updated the version-specific references to the latest, although
that isn't always necessary.  I felt like we should be more recent that
0.6.0 though.

I added a placeholder for the IBM Spectrum Conductor process proxy since
it was odd not have anything there at this point.  That can be filled in
later.  Hoping @charlieeeeeee can take a crack at that.

Fixes #365