Update kubernetes kernelspecs and launchers with initialization mode
One of the changes that occurred while kubernetes was on a separate branch
was the introduction of spark-context-initialization-mode.  These changes
include that update as well as the following found during testing:

1. Spark scala kubernetes kernels must create their spark context using the
eager mode.  For whatever reason, the Toree lazy initialization change does
not work via the spark-on-k8s fork.  However, eager works fine.  We can revist
once we switch to the Spark 2.3 code.

2. There weren't any Toree assembly files in the enterprise-gateway image. As
a result, the code that allows us to avoid pinning to specific versions doesn't
work because it requires that the assembly be present.  The Dockerfile has been
updated accordingly.

3. Added the changes from the other scala kernels that checks for the existence
of the toree assembly and launcher jar.

4. Updated handle_timeout() in processproxy.py since it wasn't using the changes
made to throw better error messages.

Fixes #366